[16.0][REF] rename web_attachment_classifier → web_attachment_document_type
CONTEXT.md flagged "classifier" as implementation-level jargon to
avoid; the module name itself violated that rule. Rename to the
formal, model-anchored name across every layer so future readers
land on consistent language.

Folder + files renamed via `git mv`:
- `web_attachment_classifier/` → `web_attachment_document_type/`
- `models/attachment_classifier_model_config.py` →
  `models/ir_attachment_document_type_config.py`
- `static/src/components/attachment_classifier_dialog.{js,xml}` →
  `attachment_document_type_dialog.{js,xml}`
- `static/src/scss/attachment_classifier.scss` →
  `attachment_document_type.scss`

Identifier renames:
- Manifest name: "Web Attachment Classifier" → "Web Attachment
  Document Type"
- Model tech name: `attachment.classifier.model.config` →
  `ir.attachment.document.type.config` (parallels sibling models
  under the `ir.attachment.document.type.*` namespace)
- Python class: `AttachmentClassifierModelConfig` →
  `IrAttachmentDocumentTypeConfig`
- JS class + template: `AttachmentClassifierDialog` →
  `AttachmentDocumentTypeDialog`
- JS instance vars: `_classifierDialogService` →
  `_doctypeDialogService`, `_classifierOrm` → `_doctypeOrm`,
  `classifierState` → `doctypeState`
- SCSS classes: `.o_classifier_*` → `.o_attachment_doctype_*`
- `patch()` key: `web_attachment_classifier.Many2ManyBinaryField`
  → `web_attachment_document_type.Many2ManyBinaryField`

Consumer updates (data XML + manifest depends):
- `kris_project` and `purchase_request_kmitl` depend on the new
  module name; data XML records point to
  `ir.attachment.document.type.config` for the Config model.

Prose cleanup: "classifier UI", "classifier-enabled", "the
classifier flow" reworded to "doctype UI", "models with doctype
mappings", "the attach + doctype flow". ADR-0001 kept as-is because
it describes a historical (superseded) design that literally used
"Classifier" in its proposed factory name; rewriting it would
rewrite history. The `_Avoid_: classifier` language rule in
CONTEXT.md is preserved — now honoured by the module name itself.

Added `# noqa: F401` markers on side-effect imports in module
`__init__.py` files (ruff --fix was reporting F401 because the
renamed paths counted as new to git).

No functional change. No DB migration needed (branch not yet
merged; dev DBs get reinstalled).
