+++ kris_project/__manifest__.py
@@ -13,7 +13,7 @@
         "tracking_manager",
         "account_analytic_kmitl",
         "hr_employee_name_detail_kmitl",
-        "web_attachment_classifier",
+        "web_attachment_document_type",
     ],
     "demo": [
         "data/kris_project_demo.xml",
