+++ purchase_request_kmitl/data/attachment_document_type.xml
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8" ?>
 <odoo noupdate="1">
 
     <!-- Document type master records -->
@@ -17,8 +17,8 @@
 
     <!-- Model configuration for purchase.request -->
 
-    <record id="config_purchase_request" model="attachment.classifier.model.config">
-        <field name="res_model_id" ref="purchase_request.model_purchase_request"/>
+    <record id="config_purchase_request" model="ir.attachment.document.type.config">
+        <field name="res_model_id" ref="purchase_request.model_purchase_request" />
         <field
             name="line_ids"
             eval="[(5, 0, 0),
