--- web_attachment_classifier/__init__.py
@@ -1,1 +1,0 @@
-from . import models
