--- web_attachment_classifier/models/__init__.py
@@ -1,4 +1,0 @@
-from . import ir_attachment
-from . import ir_attachment_document_type
-from . import attachment_classifier_model_config
-from . import ir_attachment_document_type_rel
