+++ web_attachment_document_type/__init__.py
@@ -1,0 +1,1 @@
+from . import models  # noqa: F401
