+++ web_attachment_document_type/models/__init__.py
@@ -1,0 +1,4 @@
+from . import ir_attachment  # noqa: F401
+from . import ir_attachment_document_type  # noqa: F401
+from . import ir_attachment_document_type_config  # noqa: F401
+from . import ir_attachment_document_type_rel  # noqa: F401
