+++ web_attachment_document_type/models/ir_attachment.py
