+++ web_attachment_document_type/models/ir_attachment_document_type.py
