+++ web_attachment_document_type/models/ir_attachment_document_type_rel.py
@@ -2,10 +2,10 @@ from odoo import fields, models
 
 
 class IrAttachmentDocumentTypeRel(models.Model):
-    """Line inside an `attachment.classifier.model.config` — one doctype
+    """Line inside an `ir.attachment.document.type.config` — one doctype
     on the config's model, at a given `sequence`.
 
-    The classifier widget resolves the doctype list for the current form
+    The patched many2many_binary widget resolves the doctype list for the current form
     by filtering rows on `res_model_name` (stored + indexed related
     field, sourced from `config_id.res_model_name`) and ordering by
     `sequence`. That keeps the widget's query a plain SQL scan of a
@@ -17,7 +17,7 @@ class IrAttachmentDocumentTypeRel(models.Model):
     _order = "sequence, id"
 
     config_id = fields.Many2one(
-        comodel_name="attachment.classifier.model.config",
+        comodel_name="ir.attachment.document.type.config",
         string="Model Config",
         required=True,
         ondelete="cascade",
