+++ web_attachment_document_type/static/src/patches/mail_attachment_card_patch.xml
@@ -2,10 +2,7 @@
 <templates xml:space="preserve">
 
     <t t-inherit="mail.AttachmentCard" t-inherit-mode="extension" owl="1">
-        <xpath
-            expr="//div[hasclass('o_AttachmentCard_details')]"
-            position="inside"
-        >
+        <xpath expr="//div[hasclass('o_AttachmentCard_details')]" position="inside">
             <small
                 t-if="attachmentCard.attachment.documentTypeName"
                 class="o_AttachmentCard_doctype badge text-bg-secondary mt-1 align-self-start"
