+++ web_attachment_document_type/static/src/patches/mail_attachment_patch.js
