+++ web_attachment_document_type/static/src/patches/many2many_binary_field_patch.xml
@@ -5,7 +5,7 @@
         <xpath expr="//div[@aria-atomic='true']" position="attributes">
             <attribute
                 name="t-attf-class"
-            >oe_fileupload {{props.className ? props.className : ''}} {{ classifierState.isDraggingOver ? 'o_classifier_dragging_over' : '' }}</attribute>
+            >oe_fileupload {{props.className ? props.className : ''}} {{ doctypeState.isDraggingOver ? 'o_attachment_doctype_dragging_over' : '' }}</attribute>
             <attribute name="t-on-dragenter.prevent">onDragEnter</attribute>
             <attribute name="t-on-dragover.prevent">onDragOver</attribute>
             <attribute name="t-on-dragleave.prevent">onDragLeave</attribute>
@@ -13,14 +13,14 @@
         </xpath>
         <xpath expr="//div[@aria-atomic='true']" position="inside">
             <div
-                t-if="classifierState.isDraggingOver"
-                class="o_classifier_dropzone_overlay"
+                t-if="doctypeState.isDraggingOver"
+                class="o_attachment_doctype_dropzone_overlay"
             >
                 <span class="fa fa-cloud-upload me-2" />Drop files to attach
             </div>
             <div
-                t-if="classifierState.uploading"
-                class="o_classifier_uploading_overlay"
+                t-if="doctypeState.uploading"
+                class="o_attachment_doctype_uploading_overlay"
             >
                 <span class="fa fa-spin fa-spinner me-2" />Uploading...
             </div>
@@ -39,18 +39,12 @@
             <attribute name="download" />
             <attribute name="target">_blank</attribute>
         </xpath>
-        <xpath
-            expr="//div[hasclass('caption') and hasclass('small')]"
-            position="after"
-        >
-            <div
-                t-if="classifierState.options.length"
-                class="caption small text-muted"
-            >
+        <xpath expr="//div[hasclass('caption') and hasclass('small')]" position="after">
+            <div t-if="doctypeState.options.length" class="caption small text-muted">
                 <span>Document Type:</span>
                 <span
                     t-if="file.document_type_id"
-                    t-attf-class="badge text-bg-primary ms-1 {{ editable ? 'o_classifier_badge_clickable' : '' }}"
+                    t-attf-class="badge text-bg-primary ms-1 {{ editable ? 'o_attachment_doctype_badge_clickable' : '' }}"
                     t-att-role="editable ? 'button' : null"
                     t-att-title="editable ? 'Change Document Type' : null"
                     t-on-click="() => this.onBadgeClick(file.id, file.document_type_id[0])"
@@ -58,7 +52,7 @@
                 />
                 <span
                     t-else=""
-                    t-attf-class="badge text-bg-warning ms-1 {{ editable ? 'o_classifier_badge_clickable' : '' }}"
+                    t-attf-class="badge text-bg-warning ms-1 {{ editable ? 'o_attachment_doctype_badge_clickable' : '' }}"
                     t-att-role="editable ? 'button' : null"
                     t-att-title="editable ? 'Set Document Type' : null"
                     t-on-click="() => this.onBadgeClick(file.id, null)"
