+++ web_attachment_document_type/static/src/scss/attachment_document_type.scss
@@ -1,7 +1,7 @@
 .oe_fileupload {
     position: relative;
 
-    &.o_classifier_dragging_over {
+    &.o_attachment_doctype_dragging_over {
         outline: 2px dashed var(--bs-primary, #3b82f6);
         outline-offset: -4px;
         background-color: rgba(59, 130, 246, 0.06);
@@ -9,7 +9,7 @@
     }
 }
 
-.o_classifier_uploading_overlay {
+.o_attachment_doctype_uploading_overlay {
     position: absolute;
     inset: 0;
     display: flex;
@@ -23,7 +23,7 @@
     pointer-events: none;
 }
 
-.o_classifier_dropzone_overlay {
+.o_attachment_doctype_dropzone_overlay {
     position: absolute;
     inset: 0;
     display: flex;
@@ -37,18 +37,18 @@
     border-radius: 4px;
 }
 
-.o_classifier_dialog_dropzone {
+.o_attachment_doctype_dialog_dropzone {
     border-style: dashed !important;
     border-color: var(--bs-border-color, #dee2e6) !important;
     transition: background-color 120ms ease, border-color 120ms ease;
 
-    &.o_classifier_dragging_over {
+    &.o_attachment_doctype_dragging_over {
         border-color: var(--bs-primary, #3b82f6) !important;
         background-color: rgba(59, 130, 246, 0.06);
     }
 }
 
-.o_classifier_badge_clickable {
+.o_attachment_doctype_badge_clickable {
     cursor: pointer;
     transition: filter 100ms ease, transform 100ms ease;
